Revert "adding like functionality and edit post method"
