+++ flask_app/controllers/users.py
@@ -4,7 +4,6 @@ from flask_app import app
 from flask_app.models.user import User
 from flask_app.models.post import Post
 
-
 from flask_bcrypt import Bcrypt
 bcrypt = Bcrypt(app)
 
@@ -57,8 +56,9 @@ def dashboard():
         'id':session['user_id']
     }
     user = User.get_user_by_id(data)  #grabs login information using login id from sessions
-    posts = Post.all_posts_with_likes_and_users() #grabs all posts with their user information and their cheer information
+    posts = Post.allPostsWithUserInfo() #grabs all posts with their user information and their cheer information
     friends_suggest = User.allFriendSuggestions(data)# for now showing all friends not friends with (no mutuals)
+    
     return render_template('dashboard.html' , user = user, posts = posts, friends_suggest = friends_suggest) #connects html and brings in variables into html
 
 @app.route('/addfriend/<int:id>')
@@ -107,15 +107,3 @@ def posts():
 def logout():
     session.clear()
     return redirect('/')
-
-@app.route("/like/<int:id>")
-def like_action(id):
-    if 'user_id' not in session:
-        flash("You must be signed in to view this page.")
-        return redirect('/logout')
-    data = {
-        "p_id" : id,
-        "u_id" : session['user_id']
-    }
-    Post.like_or_unlike_post(data)
-    return redirect("/dashboard")
