+++ flask_app/templates/dashboard.html
@@ -36,7 +36,6 @@
                     </div>
                     <div>
 
-
                         <!-- CREATE POST FORM -->
                         <form action="/create/post" method='post' class="bg-comment p-5 my-3">
                             {% with messages = get_flashed_messages(category_filter=['post']) %}
